AI/requirements.txt
Flask==2.1.2
requests==2.28.1
Werkzeug==2.0.3
python-dotenv==0.21.0